feat: removed polyglot in favor of treesitter and other improvements
CoC is great but treesitter is better. In particular, it has tagged
template injections in JavaScript and TypeScript. Consider the code
```ts
const myStyles = css`
  .thing {
    background: blue;
  }
`
```
With treesitter, the code inside the backticks will be highlighted as
CSS. It also appears to work for `tsx`.

I also introduced telescope, which has served as a better form of
project navigation for me. I've tried to improve the file explorer, and
to put a project directory tree on the screen (split to the right), but
both approaches have failed me. File exploration is inherently slow, and
I prefer not to waste space on my screen.